+++ server/src/modules/organization/service.ts
@@ -8,41 +8,8 @@ import { InternalServerError } from '../../shared/helpers/error';
 import { logger, logIfNotTestEnv } from '../../shared/helpers/logger';
 import DataService from '../data/data.service';
 
+import { OrganisationInfo, OrganisationResponse, ParsedOrganisation } from './organization.types';
 import { DELETE_ORGANIZATIONS, GET_ORGANIZATIONS, INSERT_ORGANIZATIONS } from './queries.gql';
-
-interface OrganisationResponse {
-	status: string;
-	description: string;
-	data: OrganisationInfo[];
-}
-
-export interface OrganisationContactInfo {
-	phone?: string;
-	website?: string;
-	email?: string;
-	logoUrl?: string;
-	form_url?: string;
-}
-
-export interface OrganisationInfo {
-	or_id: string;
-	cp_name: string;
-	category?: string;
-	sector?: string;
-	cp_name_catpro?: string;
-	description?: string;
-	contact_information: OrganisationContactInfo;
-	accountmanager?: string;
-}
-
-export interface ParsedOrganisation {
-	or_id: string;
-	name: string;
-	logo_url: string | null;
-	description: string | null;
-	website: string | null;
-	data: OrganisationInfo;
-}
 
 export default class OrganisationService {
 	public static async initialize() {
@@ -56,7 +23,7 @@ export default class OrganisationService {
 			if (process.env.NODE_ENV !== 'test') {
 				/* istanbul ignore next */
 				cron.schedule('0 0 04 * * *', async () => {
-					await OrganisationService.initialize();
+					await OrganisationService.updateOrganisationsCache();
 				}).start();
 			}
